Performance fix: ask for commits for all ref changes at once
Instead of asking for commits for each ref change separately, we now
throw all ref changes to git and let it compute list of commits to
inspect. This saves a lot of time when multiple branches are pushed
(by doing one external call instead of one per branch) and saves Java
time and memory by avoiding computing common set in code.

As a positive side-effect, it will reduce amount of commits inspected
in case a new branch is pushed together with an already existing branch.
Previously we would inspect every commit from the new branch, now we
skip commits that are also available from the already existing branch.